fix: update writing CA bundle files (#115)
If there is a crash or power outage while writing a file
it can be left in a corrupt state.

Updates are instead written to a new file and once complete
the new file is renamed to replace the original file.

This reduces the window where a file update could be lost
and reduces the chance that a file is corrupted.

Using a journaling file system can also help.

Signed-off-by: James Chapman <james.chapman@pionix.de>